requirements.txt
django~=3.0.5
appdirs==1.4.3
asn1crypto==0.24.0
cffi==1.12.3
cryptography==2.7
distlib==0.3.0
filelock==3.0.12
gunicorn==20.0.4
importlib-metadata==1.6.0
Jinja2==2.10.1
MarkupSafe==1.1.1
pycparser==2.19
PyYAML==5.1.2
six==1.12.0
virtualenv==20.0.18
virtualenvwrapper-win==1.2.6
zipp==3.1.0
django-heroku==0.3.1
django-crispy-forms==1.9.0
djangorestframework~=3.11.0